+++ CHANGELOG.md
@@ -1,2 +1,14 @@
-# 0.1.0
+# 0.2.0
++ Add Docker
++ Get public IP
++ Promisify getZoneId
++ Add Winston logger
++ Rename tests
++ Add cron
++ Add Sonarqube
++ Get subdomains list for zone
++ Add CI-CD
++ Fix CI-CD
+
+## 0.1.0
 + First version - forked